back-end/requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
google-api-core==2.2.2
google-api-python-client==2.29.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pafy==0.5.5
protobuf==3.19.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==3.0.4
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.27
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
youtube-dl==2021.6.6
zipp==3.6.0
